adjust cmd setting
