+++ azdev/commands.py
@@ -35,8 +35,9 @@ def load_command_table(self, _):
     with CommandGroup(self, 'verify', operation_group('pypi')) as g:
         g.command('history', 'check_history')
 
-    with CommandGroup(self, 'checker', operation_group('checker')) as g:
-        g.command('cmp-command-meta', 'cmp_command_meta')
+    with CommandGroup(self, 'break-change', operation_group('break_change')) as g:
+        g.command('command-meta export', 'export_command_meta')
+        g.command('command-meta diff', 'cmp_command_meta')
 
     with CommandGroup(self, 'cli', operation_group('pypi')) as g:
         g.command('check-versions', 'verify_versions')
