+++ azdev/params.py
@@ -11,6 +11,7 @@ from knack.arguments import ArgumentsContext, CLIArgumentType
 
 from azdev.completer import get_test_completion
 from azdev.operations.linter import linter_severity_choices
+from azdev.operations.break_change import diff_export_format_choices
 
 
 class Flag:
@@ -109,25 +110,24 @@ def load_arguments(self, _):
     with ArgumentsContext(self, 'statistics list-command-table') as c:
         c.positional('modules', modules_type)
 
-    with ArgumentsContext(self, 'statistics gen-command-meta') as c:
-        c.positional('modules', modules_type)
-        c.argument('with_help', action="store_false", help="State whether to include help message")
-        c.argument('with_example', action="store_false", help="State whether to include examples")
-        c.argument('meta_output_path', help='command meta json file path to store')
-
     with ArgumentsContext(self, 'statistics diff-command-tables') as c:
         c.argument('table_path', help='command table json file')
         c.argument('diff_table_path', help='command table json file to diff')
 
     # endregion
 
-    with ArgumentsContext(self, 'checker cmp-command-meta') as c:
+    with ArgumentsContext(self, 'break-change command-meta export') as c:
+        c.positional('modules', modules_type)
+        c.argument('with_help', action="store_false", help="State whether to include help message")
+        c.argument('with_example', action="store_false", help="State whether to include examples")
+        c.argument('meta_output_path', help='command meta json file path to store')
+
+    with ArgumentsContext(self, 'break-change command-meta diff') as c:
         c.argument('base_meta_path', required=True, help='command meta json file')
         c.argument('diff_meta_path', required=True, help='command meta json file to diff')
         c.argument('only_break', action="store_true", help='whether include non breaking changes')
-        c.argument('as_text', action="store_true", help='print diff and suggest message as text')
-        c.argument('as_obj', action="store_true", help='print diff and suggest message as detailed objects')
-        c.argument('as_tree', action="store_true", help='print diff and suggest message as detailed objects under command tree')
+        c.argument('output_type', choices=diff_export_format_choices(), default=diff_export_format_choices()[0],
+                   help='format to print diff and suggest message')
         c.argument('output_file', help='command meta diff json file path to store')
 
     with ArgumentsContext(self, 'perf') as c:
